Refactored CreatePost and moved main logic to a different component for re-usability.
