+++ src/screens/Feed/components/ModalHeader/component.js
@@ -10,19 +10,32 @@ export const ModalHeader = ({
   leftButtonTitle,
   leftButtonAction,
   leftButtonDisabled,
+  leftButtonLoading,
   rightButtonTitle,
   rightButtonAction,
   rightButtonDisabled,
+  rightButtonLoading,
 }) => (
   <View style={styles.modalHeader}>
     <View style={styles.modalButton}>
-      <HeaderButton onPress={leftButtonAction} title={leftButtonTitle} disabled={leftButtonDisabled} />
+      <HeaderButton
+        onPress={leftButtonAction}
+        title={leftButtonTitle}
+        disabled={leftButtonDisabled}
+        loading={leftButtonLoading}
+      />
     </View>
     <View style={styles.modalTitle}>
       <StyledText color="light" size="small" bold>{title}</StyledText>
     </View>
     <View style={[styles.modalButton, styles.modalRightButton]}>
-      <HeaderButton onPress={rightButtonAction} highlighted title={rightButtonTitle} disabled={rightButtonDisabled} />
+      <HeaderButton
+        onPress={rightButtonAction}
+        highlighted
+        title={rightButtonTitle}
+        disabled={rightButtonDisabled}
+        loading={rightButtonLoading}
+      />
     </View>
   </View>
 );
@@ -32,9 +45,11 @@ ModalHeader.propTypes = {
   leftButtonTitle: PropTypes.string,
   leftButtonAction: PropTypes.func,
   leftButtonDisabled: PropTypes.bool,
+  leftButtonLoading: PropTypes.bool,
   rightButtonTitle: PropTypes.string,
   rightButtonAction: PropTypes.func,
   rightButtonDisabled: PropTypes.bool,
+  rightButtonLoading: PropTypes.bool,
 };
 
 ModalHeader.defaultProps = {
@@ -42,7 +57,9 @@ ModalHeader.defaultProps = {
   leftButtonTitle: 'Cancel',
   leftButtonAction: null,
   leftButtonDisabled: false,
+  leftButtonLoading: false,
   rightButtonTitle: 'Done',
   rightButtonAction: null,
   rightButtonDisabled: false,
+  rightButtonLoading: false,
 };
