+++ src/screens/Feed/components/PostCreator/GIFImage.js
