+++ src/screens/Feed/components/PostCreator/index.js
@@ -1,0 +1,11 @@
+import { GIFImage } from './GIFImage';
+import { AdditionalButton } from './AdditionalButton';
+import { MediaItem } from './MediaItem';
+import PostCreator from './PostCreator';
+
+export {
+  GIFImage,
+  AdditionalButton,
+  MediaItem,
+  PostCreator,
+};
